+++ pentobi/Board.qml
@@ -221,7 +221,7 @@ Item {
                 text: getColumnCoord(index)
                 color: theme.colorText
                 opacity: 0.55 - 0.1 * theme.colorBackground.hslLightness
-                font { pixelSize: coordinateFontSize; preferShaping: false }
+                font.pixelSize: coordinateFontSize
                 x: mapFromGameX(index) - grabImageTarget.x
                    + (gridWidth - width) / 2
                 y: mapFromGameY(-1) - grabImageTarget.y
@@ -235,7 +235,7 @@ Item {
                 text: getColumnCoord(index)
                 color: theme.colorText
                 opacity: 0.55 - 0.1 * theme.colorBackground.hslLightness
-                font { pixelSize: coordinateFontSize; preferShaping: false }
+                font.pixelSize: coordinateFontSize
                 x: mapFromGameX(index) - grabImageTarget.x
                    + (gridWidth - width) / 2
                 y: mapFromGameY(rows) - grabImageTarget.y
@@ -249,7 +249,7 @@ Item {
                 text: index + 1
                 color: theme.colorText
                 opacity: 0.55 - 0.1 * theme.colorBackground.hslLightness
-                font { pixelSize: coordinateFontSize; preferShaping: false }
+                font.pixelSize: coordinateFontSize
                 x: mapFromGameX(isTrigon ? -1.5 : -1) - grabImageTarget.x
                    + (gridWidth - width) / 2
                 y: mapFromGameY(rows - index - 1) - grabImageTarget.y
@@ -263,7 +263,7 @@ Item {
                 text: index + 1
                 color: theme.colorText
                 opacity: 0.55 - 0.1 * theme.colorBackground.hslLightness
-                font { pixelSize: coordinateFontSize; preferShaping: false }
+                font.pixelSize: coordinateFontSize
                 x: mapFromGameX(isTrigon ? columns + 0.5 : columns)
                     - grabImageTarget.x + (gridWidth - width) / 2
                 y: mapFromGameY(rows - index - 1) - grabImageTarget.y