+++ pentobi/PieceNexos.qml
@@ -132,10 +132,7 @@ PieceBase
                 width: board.gridWidth
                 height: board.gridHeight
                 fontSizeMode: Text.Fit
-                font {
-                    pixelSize: 0.5 * board.gridHeight
-                    preferShaping: false
-                }
+                font.pixelSize: 0.5 * board.gridHeight
                 minimumPixelSize: Math.max(3, 0.3 * board.gridHeight)
                 verticalAlignment: Text.AlignVCenter
                 horizontalAlignment: Text.AlignHCenter