+++ pentobi/PieceTrigon.qml
@@ -92,10 +92,7 @@ PieceBase
                 width: board.gridWidth
                 height: board.gridHeight
                 fontSizeMode: Text.Fit
-                font {
-                    pixelSize: 0.5 * board.gridHeight
-                    preferShaping: false
-                }
+                font.pixelSize: 0.5 * board.gridHeight
                 minimumPixelSize: 5
                 verticalAlignment: Text.AlignVCenter
                 horizontalAlignment: Text.AlignHCenter